+++ puzzles/2023/5/sources/RangeMap.cpp
@@ -3,14 +3,14 @@
 
 namespace aoc_2023_5 {
 
-RangeMapSection::RangeMapSection(uint32_t dest, uint32_t src, uint32_t length)
+RangeMapSection::RangeMapSection(uint64_t dest, uint64_t src, uint64_t length)
     : mDestinationStart{dest}
     , mSourceStart{src}
     , mRangeLength{length}
 {
 }
 
-std::optional<uint32_t> RangeMapSection::get(uint32_t const key) const noexcept
+std::optional<uint64_t> RangeMapSection::get(uint64_t const key) const noexcept
 {
     if (key < mSourceStart || key > (mSourceStart + mRangeLength)) {
         return {};
@@ -19,12 +19,12 @@ std::optional<uint32_t> RangeMapSection::get(uint32_t const key) const noexcept
 }
 
 void RangeMap::addSection(
-    uint32_t const dest, uint32_t const src, uint32_t const length) noexcept
+    uint64_t const dest, uint64_t const src, uint64_t const length) noexcept
 {
     mSections.emplace_back(dest, src, length);
 }
 
-uint32_t RangeMap::get(uint32_t key) const noexcept
+uint64_t RangeMap::get(uint64_t key) const noexcept
 {
     for (auto const& section : mSections) {
         auto const value{section.get(key)};
