+++ puzzles/2023/5/sources/RangeMap.hpp
@@ -8,20 +8,20 @@ namespace aoc_2023_5 {
 
 class RangeMapSection {
 public:
-    RangeMapSection(uint32_t dest, uint32_t src, uint32_t length);
-    [[nodiscard]] std::optional<uint32_t> get(uint32_t key) const noexcept;
+    RangeMapSection(uint64_t dest, uint64_t src, uint64_t length);
+    [[nodiscard]] std::optional<uint64_t> get(uint64_t key) const noexcept;
 
 private:
-    uint32_t mDestinationStart;
-    uint32_t mSourceStart;
-    uint32_t mRangeLength;
+    uint64_t mDestinationStart;
+    uint64_t mSourceStart;
+    uint64_t mRangeLength;
 };
 
 class RangeMap {
 public:
     RangeMap() = default;
-    [[nodiscard]] uint32_t get(uint32_t key) const noexcept;
-    void addSection(uint32_t dest, uint32_t src, uint32_t length) noexcept;
+    [[nodiscard]] uint64_t get(uint64_t key) const noexcept;
+    void addSection(uint64_t dest, uint64_t src, uint64_t length) noexcept;
 
 private:
     std::vector<RangeMapSection> mSections{};
